requirements.txt
Attest==0.5.3
Fabric==1.4.0
Flask==0.8
Flask-Celery==2.4.2
Flask-Login==0.1
Flask-Mail==0.6.1
Flask-PyMongo==0.1
Flask-Script==0.3.3
Flask-Testing==0.3
Flask-fillin==0.2
Jinja2==2.6
Pygments==1.5
Werkzeug==0.8.3
amqplib==1.0.2
anyjson==0.3.1
argparse==1.2.1
blinker==1.2
celery==2.5.1
celery-with-mongodb==2.5
chardet==1.0.1
importlib==1.0.2
kombu==2.1.2
lamson==1.1
lockfile==0.9.1
lxml==2.3.4
mock==0.8.0
nose==1.1.2
ordereddict==1.1
progressbar==2.3
py==1.4.7
pycrypto==2.5
pymongo==2.1.1
pytest==2.2.3
python-daemon==1.6
python-dateutil==1.5
simplejson==2.4.0
ssh==1.7.13
wsgiref==0.1.2
